requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
django-polymorphic==3.1.0
psycopg2-binary==2.9.4
sqlparse==0.4.3
typing_extensions==4.4.0
